// packages/charm/src/types.ts
import {
  RouteProps,
  RouteComponentProps,
  match as Match,
} from 'react-router-dom';
import { HelmetData } from 'react-helmet';
import { Request, Response } from 'express';
import { History, Location } from 'history';

export interface CtxBase {
  req?: Request;
  res?: Response;
  history?: History;
  location?: Location;
  scrollToTop?: ScrollToTop;
}

export interface Ctx<P> extends CtxBase {
  match: Match<P>;
}

// TODO: possibly add unique hash + cache
// LazyLoaded component with chunk hash
export interface AsyncComponent {
  getInitialProps: (props: Ctx<any>) => any;
  load?: () => Promise<React.ReactNode>;
  getChunkName: () => string | undefined;
}

export type AsyncRouteComponentType<Props> = React.ComponentType<Props> &
  AsyncComponent;

/*
  this type handles the component type in the route config object
    * {
    *   path: "/",
    *   exact: true,
    *   component: ReactComponent <- AsyncRouteComponent
    * }
*/
export type AsyncRouteComponent<Props = any> =
  // re-exported from react-router (RouteComponentProps)
  | React.ComponentType<RouteComponentProps<Props>>
  | React.ComponentType<Props>
  // Charm Container Component (getInitialProps and ...)
  | AsyncRouteComponentType<RouteComponentProps<Props>>
  | AsyncRouteComponentType<Props>;

export interface AsyncRouteComponentState {
  Component: AsyncRouteComponent | null;
}

export interface AsyncRouteProps<Props = any> extends RouteProps {
  path?: string;
  Placeholder?: React.ComponentType<any>;
  component: AsyncRouteComponent<Props>;
  redirectTo?: string;
}

export type ScrollToTop = React.RefObject<boolean>;

// Result of getInitialProps
export type InitialData = Promise<unknown>[];

// Result of InitialData + Route Props
export interface InitialProps {
  match?: AsyncRouteProps;
  data: InitialData;
}

// <CharmData /> will send this object
export interface ServerAppState {
  quarkData: CharmClientData;
  initialData: InitialData;
}

export interface CharmClientData {
  scrollToTop: ScrollToTop;
}

// Result of Document
export interface RenderPageResult {
  html: string;
  helmet: HelmetData;
}

// TransitionBehavior
export type TransitionBehavior = 'blocking' | 'instant';

// Document.js
export interface DocumentGetInitialProps<T = RenderPageResult> {
  req: Request;
  res: Response;
  helmet: HelmetData;
  assets: Assets;
  data: ServerAppState;
  renderPage: () => Promise<T>;
  match: Match | null;
  scripts: string[];
  styles: string[];
  scrollToTop: ScrollToTop;
}

export type DocumentProps<T = RenderPageResult> = Omit<
  DocumentGetInitialProps<T>,
  'req' | 'res' | 'renderPage' | 'scrollToTop'
> &
  T;
export type CharmContext = DocumentProps;

// getAssets utility function
export interface GetAssetsParams {
  chunks: Chunks;
  route?: AsyncRouteProps<any>;
}

// ES Module type
export type Module<P> =
  | {
      default?: P;
      [x: string]: any;
    }
  | {
      exports?: P;
      [x: string]: any;
    };

// assets.json that generated by matter
export interface Assets {
  [name: string]: {
    [ext: string]: string;
  };
}

// chunks.json that generated by matter
export interface Chunks {
  [key: string]: {
    css: string[];
    js: string[];
  };
}